Fix test.py when using CPU

+++ test.py
@@ -87,9 +87,9 @@ def test_classifier(model, val_dataset, model_results, binary_classification, bi
         if binary_classification:
             prediction = sigmoid(prediction).item()
         else:
-            prediction = (exp(prediction.cpu().detach()) @ bin_centers).item()
+            prediction = (exp(prediction.float().cpu().detach()) @ bin_centers).item()
         predictions.append(prediction)
-        label = target.item() if binary_classification else (target.cpu() @ bin_centers).item()
+        label = target.item() if binary_classification else (target.float().cpu() @ bin_centers).item()
         labels.append(label)
     compute_metrics(predictions, labels, binary_classification, model_results)
     return labels
